clear code, component GenericForm really generic
